update/fix build_sphinx
Use the sphinx provided BuildDoc command rather than the one I ignorantly
hacked together.

You may not like this because it introduces a dependency of setup.py
on sphinx. There was an implied dependence before with the system call
to sphinx-build. But in that case not having sphinx would only be
noticed and cause a problem when the build_sphinx command was invoked.
Now setup.py has a first class, up front dependency on sphinx with the
import of sphinx.setup_command.BuildDoc. We can make this
traditional usage of build_sphinx a dynamic dependency by wrapping
the import with a try/except block and adding logic to define or bypass
the command settings based upon the success of the import.